don't overwrite config, instead use the credentials passed from the constructor. After killing the tunnel wait for it to die

+++ lib/testingbot/tunnel.rb
@@ -8,7 +8,6 @@ module TestingBot
       @available = false
       @errors = []
       @config = TestingBot.get_config
-      @config.add_options(opts)
       @options = default_options
       @options = @options.merge(opts)
 
@@ -19,10 +18,10 @@ module TestingBot
       @config.require_tunnel
       
       p "Starting the TestingBot Tunnel" if @options[:verbose] == true
-      @process = IO.popen("exec java -jar #{get_jar_path} #{config[:client_key]} #{config[:client_secret]} #{extra_options} 2>&1")
+      @process = IO.popen("exec java -jar #{get_jar_path} #{@options[:client_key] || @config[:client_key]} #{@options[:client_secret] || @config[:client_secret]} #{extra_options} 2>&1")
       at_exit do
         # make sure we kill the tunnel
-        Process.kill("INT", @process.pid)
+        Process.kill("INT", @process.pid) if @available == true
       end
 
       Thread.new do 
@@ -54,9 +53,10 @@ module TestingBot
 
     def stop
       raise "Can't stop tunnel, it has not been started yet" if @process.nil?
-
+      @available = false
       p "Stopping TestingBot Tunnel" if @options[:verbose] == true
       Process.kill("INT", @process.pid)
+      Process.wait
     end
 
     private
